+++ unity/Assets/QuestNav/Network/NetworkTables-CSharp.meta
