+++ unity/Assets/QuestNav/Network/NetworkTables-CSharp/Nt4SubscriptionOptions.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace NetworkTables
+namespace QuestNav.Network.NetworkTables_CSharp
 {
     public class Nt4SubscriptionOptions
     {
