+++ unity/Assets/QuestNav/UI/UIManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 84aa431546284adea2793a13d73b0c75
+timeCreated: 1742238871
