+++ unity/Assets/QuestNav/Utils/LoggerFlusher.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace QuestNav.Utils
+{
+    public class LoggerFlusher : MonoBehaviour
+    {
+        void Update()
+        {
+            QueuedLogger.Flush();
+        }
+    }
+}
