+++ unity/Assets/QuestNav/Utils/VectorExtensions.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c4493859a6794216a56d42a82c18d376
+timeCreated: 1742238908
