--- unity/Assets/Robot.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5498358051692774eac48dc5801b2cc3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
