--- unity/Assets/Robot/LoggerFlusher.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class LoggerFlusher : MonoBehaviour
-{
-    void Update()
-    {
-        QueuedLogger.Flush();
-    }
-}
