--- unity/Assets/Robot/QuestNav.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb42b4655facea54e87c326fcc93986a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
